Comenzando sesiones

+++ elimina_liga.php
@@ -1,28 +1,36 @@
 <?php
 require_once("includes/functions.php");
 require_once("db.php");
-head("Elimina");
+session_start();
 
-if (isset($_GET['liga'])) {
-    $ligaId = $_GET['liga'];
-    if (isset($_GET['si']) == "on") {
-        $sql = "delete from partido where liga_id = $ligaId";
-        mysqli_query($con, $sql);
-        $sql = "delete from equipo where liga_id = $ligaId";
-        mysqli_query($con, $sql);
-        $sql = "delete from liga where id = $ligaId";
-        mysqli_query($con, $sql);
+if ($_SESSION['username']) {
+    head("Elimina");
+
+    if (isset($_GET['liga'])) {
+        $ligaId = $_GET['liga'];
+        if (isset($_GET['si']) == "on") {
+            $sql = "delete from partido where liga_id = $ligaId";
+            mysqli_query($con, $sql);
+            $sql = "delete from equipo where liga_id = $ligaId";
+            mysqli_query($con, $sql);
+            $sql = "delete from liga where id = $ligaId";
+            mysqli_query($con, $sql);
+            header("Location: index.php");
+        }
+    } else {
         header("Location: index.php");
     }
+
+    ?>
+
+    <form action="elimina_liga.php" method="get">
+        <input type="hidden" name="liga" value="<?= $ligaId ?>">
+        <label for="si">¿Está seguro?</label><input type="checkbox" name="si" id="si">
+        <input type="submit" value="Envia">
+    </form>
+
+<?php
 } else {
-    header("Location: index.php");
+header("Location: index.php");
 }
-
 ?>
-
-<form action="elimina_liga.php" method="get">
-    <input type="hidden" name="liga" value="<?= $ligaId ?>">
-    <label for="si">¿Está seguro?</label><input type="checkbox" name="si" id="si">
-    <input type="submit" value="Envia">
-</form>
-
